fix bug of line click on canvas
bug caused because canvas height wasn't 500px
as defined.
it was 615px and the reason for it is that editor container has flex-grow 1 in order to stick footer to the bottom of the page.
solved it by set max-height of canvas to 500px, and also set settings container align items center so the items in it will look in center.