refactor render + set bg img + download meme
now only when gallery img is clicked,
we wait to the bg img to load and then draw on canvas.
in that way, we don't need to load bg img on every render, and the download don't need to wait to render to finsuh,
because render is synchrounous and do not wait to any load.

IT IS IMPORTANT TO DO THINGS ONLY WHEN NEEDED.
LOADING BG IMG EVERY RENDER IS BAD PERFORMENCE.